% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGNAF.R
\name{buildGNAF}
\alias{buildGNAF}
\title{buildGNAF}
\usage{
buildGNAF(gnafPath, writePath = NULL, Overwrite = F)
}
\arguments{
\item{path}{The path to GNAF data directory}
}
\value{
boolean
}
\description{
Builds the SQLite database that will house the GNAF data used to geocode
}
\examples{
buildGNAF('~/GNAF')
buildGNAF('~/Downloads/MAY16_GNAF+EULA_PipeSeparatedValue_20160523140820.zip'
           ,writePath = '~/Downloads'
           ,Overwrite = T)
}

