% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGNAF.R
\name{insertData}
\alias{insertData}
\title{insert Data}
\usage{
insertData(GNAFDirectory)
}
\arguments{
\item{GNAFDirectory}{}
}
\value{
T
}
\description{
insert Data
}
\examples{
#'insertData('~/Downloads/GNAF')
}

