% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGNAF.R
\name{createTables}
\alias{createTables}
\title{createTables}
\usage{
createTables(GNAFDirectory, overwrite)
}
\arguments{
\item{GNAFDirectory}{}
}
\value{
T
}
\description{
Creates GNAF table structure
}
\examples{
#' createTables('~/Downloads/GNAF')
}

